refactoring: move public constant/types to enry.go
Also re-arrange public API functions to have the same order
as they are in tests.
